requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.3.0
charset-normalizer==2.0.12
click==8.1.3
discord.py @ git+https://github.com/Rapptz/discord.py.git@06df0d7931385a5b8efc1dc6bea79cc389b22808
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
lxml==4.8.0
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.2.0
uvloop==0.16.0
yarl==1.7.2
